First code commit
This is about two days worth of work being commited for the first time, (new to github).

Will be commiting at every milestone from now on, or at end of work.

File stared with the group work myself and Andrew started in class,  whilst using the callibrate sensor from he line follower example.

The GUI is a basic button layout with a consol for the user.